pyproject.toml
[tool.poetry]
name = "dboard"
version = "0.1.0"
description = ""
authors = ["Marcelo Félix"]

[tool.poetry.dependencies]
python = "^3.8"
rich = "^12.5.1"
PyYAML = "^6.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
